+++ Source/Public/Set-VsteamWorkItemPage.ps1
@@ -1,5 +1,5 @@
-function Set-VsteamWorkItemPage {
-   [CmdletBinding(SupportsShouldProcess=$true)]
+function Set-VSTeamWorkItemPage {
+   [CmdletBinding(SupportsShouldProcess=$true,ConfirmImpact='High')]
    param(
       [parameter(ValueFromPipelineByPropertyName=$true)]
       [vsteam_lib.ProcessTemplateValidateAttribute()]
@@ -10,51 +10,52 @@ function Set-VsteamWorkItemPage {
       [ArgumentCompleter([vsteam_lib.WorkItemTypeCompleter])]
       $WorkItemType,
 
-      [parameter(Mandatory = $true, Position=1)]
+      [Parameter(ValueFromPipelineByPropertyName=$true)]
+      [ArgumentCompleter([vsteam_lib.PageCompleter])]
       [Alias('Name','PageLabel')]
       $Label,
 
-      [int]$Order,
+      [Parameter(ParameterSetName='Label',Mandatory=$true)]
+      [Parameter(ParameterSetName='Both',Mandatory=$true)]
+      [string]$Newlabel,
 
-      [ValidateSet('custom','attachments','history','links')]
-      [String]$PageType = 'custom',
+      [Parameter(ParameterSetName='Order',Mandatory=$true)]
+      [Parameter(ParameterSetName='Both',Mandatory=$true)]
+      [int]$Order,
 
       $Sections,
 
-      [switch]
-      $Force
+      [switch]$Force
    )
    process {
-      #This is designed to allow multiple pages to be added, and/or multiple WorkItemTypes to be modifed
-      if ($Label.count -gt 1 -and ($order -or $Sections)) {throw "Can't process multiple pages when Order and/or sections are specified."  ; return}
-         #Get the workitem type(s) we are updating. If we get a system one, make it an inherited one first before adding the page.
-         $wit = Get-VSTeamWorkItemType -ProcessTemplate $ProcessTemplate -WorkItemType  $WorkItemType
-         foreach ($w in $wit) {
-            if ($w.customization -eq 'system') {
-               $url  = ($w.url -replace '/workItemTypes/.*$', '/workItemTypes?api-version=') +  (_getApiVersion Processes)
-               $body = @{
-                     color        = $w.color
-                     description  = $w.description
-                     icon         = $w.icon
-                     inheritsFrom = $w.referenceName
-                     isDisabled   = $w.isDisabled
-                     name         = $w.name
-               }
-               $w = _callAPI -Url $url -method Post -body (ConvertTo-Json $body)
-            }
-            $url= $w.url + "/layout/pages?api-version=" + (_getApiVersion Processes)
-            foreach ($l in $Label) {
-               $body = @{
-                        label        = $l
-                        pageType     = $PageType
-                        visible      = $true
-               }
-               if ($PSBoundParameters.ContainsKey('Order'))    {$body['order']=$Order}
-               if ($PSBoundParameters.ContainsKey('Sections')) {$body['sections']=$Sections}
-##PATCH https://dev.azure.com/{organization}/_apis/work/processes/{processId}/workItemTypes/{witRefName}/layout/pages?api-version=5.1-preview.1
-               $resp = _callAPI -method Post -Url $url -body (ConvertTo-Json $body)
+      #Get the workitem type(s) we are updating. If we get a system one, make it an inherited one first
+      $wit = Get-VSTeamWorkItemType -ProcessTemplate $ProcessTemplate -WorkItemType $WorkItemType -Expand layout |
+             Where-object {$_.layout.pages.where({$_.label -like $Label -and -not $_.locked })}
+      if (-not $wit) {
+         Write-Warning "Could not find an unlocked page matching '$Label' for WorkItemType '$WorkItemType'."
+         return
+      }
+      $wit = $wit | Unlock-VsteamWorkItemType -Force:$Force -Expand layout
+      foreach ($w in $wit) {
+         $url= $w.url + "/layout/pages?api-version=" + (_getApiVersion Processes)
+         $page = $w.layout.pages.where({$_.label -like $Label})
+         if ($page.count -gt 1) {
+            Write-Warning "'$label' Matches more than one page on $($w.name)."
+            continue
+         }
+         $body = @{id = $page.id}
+         if ($PSBoundParameters.ContainsKey('Newlabel')) {$body['label']=$Newlabel}
+         else {$body['label'] = $page.label}
+         if ($PSBoundParameters.ContainsKey('Order'))    {$body['order']=$Order}
+         if ($PSBoundParameters.ContainsKey('Sections')) {$body['sections']=$Sections}
+         if ($Force -or $PSCmdlet.ShouldProcess("'$($page.Label)' of '$($w.name)'",'Update Page')) {
+               #Call the REST API
+               $resp = _callAPI -method Patch -Url $url -body (ConvertTo-Json $body)
+
+               # Apply a Type Name so we can use custom format view and custom type extensions
+               # and add members to make it easier if piped into something which takes values by property name
                $resp.psobject.TypeNames.Insert(0,'vsteam_lib.Workitempage')
-               Add-Member -InputObject $resp -MemberType NoteProperty  -Name PageLabel       -Value $l
+               Add-Member -InputObject $resp -MemberType AliasProperty -Name PageLabel       -Value 'label'
                Add-Member -InputObject $resp -MemberType NoteProperty  -Name WorkItemType    -Value $w.name
                Add-Member -InputObject $resp -MemberType NoteProperty  -Name ProcessTemplate -Value $ProcessTemplate
 
@@ -62,4 +63,3 @@ function Set-VsteamWorkItemPage {
             }
 }   }
 }
-